+++ bin/c-d-image-prune.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const program = require('commander');
+const { exec } = require('shelljs');
+
+// The basic program, which uses sub-commands.
+program.parse(process.argv);
+
+exec('docker image prune');