+++ bin/lib/c-ts.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const app = '/Applications/Tailscale.app/Contents/MacOS/Tailscale';
+
+module.exports = {
+    app,
+};